# services/sql-assessment-service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SQL Assessment Service API",
    "version": "1.0.0",
    "description": "REST API for SQL exercise generation, student query grading, and natural-language description of SQL queries in an educational context.",
    "contact": {
      "name": "ALADIN Project"
    }
  },
...
  "paths": {
...
    "/api/generation/generate": {
      "get": {
        "summary": "Generate a random SQL task",
        "description": "Generates a random SQL SELECT query based on the supplied task configuration (join depth, aggregation, predicates, etc.) and produces up to five natural-language descriptions of the task using different generation strategies (template, LLM entity-relationship, LLM schema-based, LLM creative, hybrid). The database must have been previously registered via POST /api/database/analyze-database.\n",
        "tags": [
          "Generation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task generated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid connection info, unregistered database, or invalid task configuration.\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Query or description generation failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error description (localised).",
            "example": "Unable to connect to the database."
          },
          "code": {
            "type": "string",
            "description": "Optional machine-readable error code.",
            "example": "NON_SELECT"
          }
        }
      },
...
      "PostgresConnectionInfo": {
        "type": "object",
        "required": [
          "type",
          "host",
          "port",
          "username",
          "password",
          "database",
          "schema"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "postgres"
            ],
            "example": "postgres"
          },
          "host": {
            "type": "string",
            "example": "localhost"
          },
          "port": {
            "type": "number",
            "example": 5432
          },
          "username": {
            "type": "string",
            "example": "myuser"
          },
          "password": {
            "type": "string",
            "example": "mypassword"
          },
          "database": {
            "type": "string",
            "example": "mydb"
          },
          "schema": {
            "type": "string",
            "example": "public"
          }
        }
      },
...
      "TaskConfiguration": {
        "type": "object",
        "required": [
          "aggregation",
          "orderby",
          "joinDepth",
          "joinTypes",
          "predicateCount",
          "groupby",
          "having",
          "columnCount",
          "operationTypes"
        ],
        "properties": {
          "aggregation": {
            "type": "boolean",
            "description": "Whether to include aggregate functions (COUNT, SUM, …)."
          },
          "orderby": {
            "type": "boolean",
            "description": "Whether to include an ORDER BY clause."
          },
          "joinDepth": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of JOIN operations to include."
          },
          "joinTypes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "INNER JOIN",
                "LEFT JOIN",
                "RIGHT JOIN",
                "FULL JOIN",
                "CROSS JOIN",
                "SELF JOIN"
              ]
            }
          },
          "predicateCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of WHERE predicates to generate."
          },
          "groupby": {
            "type": "boolean",
            "description": "Whether to include a GROUP BY clause."
          },
          "having": {
            "type": "boolean",
            "description": "Whether to include a HAVING clause (requires groupby=true)."
          },
          "columnCount": {
            "type": "integer",
            "minimum": 1,
            "description": "Number of columns to select."
          },
          "operationTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Predicate operation types to use (e.g. \"=\", \">\", \"LIKE\")."
          }
        }
      },
      "GenerateTaskRequest": {
        "type": "object",
        "required": [
          "connectionInfo",
          "taskConfiguration"
        ],
        "properties": {
          "connectionInfo": {
            "$ref": "#/components/schemas/PostgresConnectionInfo"
          },
          "taskConfiguration": {
            "$ref": "#/components/schemas/TaskConfiguration"
          },
          "languageCode": {
            "type": "string",
            "example": "en"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "required": [
          "query",
          "templateBasedDescription",
          "gptEntityRelationshipDescription",
          "gptSchemaBasedDescription",
          "hybridDescription"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "The generated SQL SELECT query.",
            "example": "SELECT name FROM customers WHERE age > 30"
          },
          "templateBasedDescription": {
            "type": "string",
            "description": "AST-template-generated natural-language description."
          },
          "gptEntityRelationshipDescription": {
            "type": "string",
            "description": "LLM-generated description using entity-relationship context (multi-step)."
          },
          "gptSchemaBasedDescription": {
            "type": "string",
            "description": "LLM-generated description using the raw schema (default GPT option)."
          },
          "hybridDescription": {
            "type": "string",
            "description": "Hybrid (template + LLM) natural-language description."
          },
          "gptCreativeDescription": {
            "type": "string",
            "description": "LLM-generated creative description (temperature 0.7). May be absent."
          }
        }
      },
...
    }
  }
}